requirements.txt
# Core dependencies for MH-C2C research
openai>=0.28.0
tiktoken>=0.5.0
tenacity>=8.0.0
python-dotenv>=1.0.0

# Optional dependencies for extended functionality
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910